refactor: opt blendshape animation
