+++ packages/core/src/mesh/BlendShapeManager.ts
@@ -1,4 +1,5 @@
 import { Engine } from "../Engine";
+import { VertexElement, VertexElementFormat } from "../graphic";
 import { Texture2DArray, TextureFilterMode, TextureFormat } from "../texture";
 import { UpdateFlag } from "../UpdateFlag";
 import { BlendShape } from "./BlendShape";
@@ -17,6 +18,9 @@ export class BlendShapeManager {
   _blendShapes: BlendShape[] = [];
   /** @internal */
   _blendShapeUpdateFlags: UpdateFlag[] = [];
+
+  /** @internal */
+  _usingTextureStoreData: boolean = false;
   /** @internal */
   _blendShapeDataTexture: Texture2DArray;
 
@@ -55,7 +59,98 @@ export class BlendShapeManager {
   /**
    * @internal
    */
-  _updateBlendShapeDataTexture(): void {
+  _supplementalVertexElement(vertexElements: VertexElement[], offset: number): number {
+    let supplementalElementCount = 0;
+    for (let i = 0, n = this._blendShapes.length; i < n; i++) {
+      vertexElements.push(new VertexElement(`POSITION_BS${i}`, offset, VertexElementFormat.Vector3, 0));
+      offset += 12;
+      supplementalElementCount += 3;
+      if (this._useBlendShapeNormal) {
+        vertexElements.push(new VertexElement(`NORMAL_BS${i}`, offset, VertexElementFormat.Vector3, 0));
+        offset += 12;
+        supplementalElementCount += 3;
+      }
+      if (this._useBlendShapeTangent) {
+        vertexElements.push(new VertexElement(`TANGENT_BS${i}`, offset, VertexElementFormat.Vector3, 0));
+        supplementalElementCount += 3;
+      }
+    }
+    return supplementalElementCount;
+  }
+
+  /**
+   * @internal
+   */
+  _updateDataToVertices(vertices: Float32Array, offset: number, vertexCount: number, elementCount: number): void {
+    const blendShapes = this._blendShapes;
+    const blendShapeUpdateFlags = this._blendShapeUpdateFlags;
+    const blendShapeCount = Math.min(blendShapes.length, 4);
+
+    if (!this._usingTextureStoreData) {
+      for (let i = 0; i < blendShapeCount; i++) {
+        const blendShapeUpdateFlag = blendShapeUpdateFlags[i];
+        if (blendShapeUpdateFlag.flag) {
+          const blendShape = blendShapes[i];
+          const { frames } = blendShape;
+          const frameCount = frames.length;
+          const endFrame = frames[frameCount - 1];
+          if (frameCount > 0 && endFrame.deltaPositions.length !== vertexCount) {
+            throw "BlendShape frame deltaPositions length must same with mesh vertexCount.";
+          }
+
+          const { deltaPositions } = endFrame;
+          for (let j = 0; j < vertexCount; j++) {
+            const start = elementCount * j + offset;
+            const deltaPosition = deltaPositions[j];
+            if (deltaPosition) {
+              vertices[start] = deltaPosition.x;
+              vertices[start + 1] = deltaPosition.y;
+              vertices[start + 2] = deltaPosition.z;
+            }
+          }
+          offset += 3;
+
+          if (this._useBlendShapeNormal) {
+            const { deltaNormals } = endFrame;
+            if (deltaNormals) {
+              for (let j = 0; j < vertexCount; j++) {
+                const start = elementCount * j + offset;
+                const deltaNormal = deltaNormals[j];
+                if (deltaNormal) {
+                  vertices[start] = deltaNormal.x;
+                  vertices[start + 1] = deltaNormal.y;
+                  vertices[start + 2] = deltaNormal.z;
+                }
+              }
+            }
+            offset += 3;
+          }
+
+          if (this._useBlendShapeTangent) {
+            const { deltaTangents } = endFrame;
+            if (deltaTangents) {
+              for (let j = 0; j < vertexCount; j++) {
+                const start = elementCount * j + offset;
+                const deltaTangent = deltaTangents[j];
+                if (deltaTangent) {
+                  vertices[start] = deltaTangent.x;
+                  vertices[start + 1] = deltaTangent.y;
+                  vertices[start + 2] = deltaTangent.z;
+                }
+              }
+            }
+            offset += 3;
+          }
+          blendShapeUpdateFlag.flag = false;
+        }
+      }
+    }
+  }
+
+  /**
+   * @internal
+   */
+  _updateDataToTexture(): void {
     const vertexCount = 0;
     const maxTextureSize = 0;
 
@@ -127,6 +222,9 @@ export class BlendShapeManager {
     for (let i = 0, n = blendShapes.length; i < n; i++) {}
   }
 
+  /**
+   * @internal
+   */
   _releaseCache(): void {
     const blendShapeUpdateFlags = this._blendShapeUpdateFlags;
     for (let i = 0, n = blendShapeUpdateFlags.length; i < n; i++) {
